Change payment commission type to float

+++ cache/src/main/kotlin/com/kg/gettransfer/cache/model/ConfigsCached.kt
@@ -19,7 +19,7 @@ data class ConfigsCached(
     @ColumnInfo(name = ConfigsEntity.AVAILABLE_LOCALES) val availableLocales: LocaleCachedList,
     @ColumnInfo(name = ConfigsEntity.PREFERRED_LOCALE) val preferredLocale: String,
     @Embedded(prefix = ConfigsEntity.CARD_GATEWAYS) val cardGateways: CardGatewaysCached,
-    @ColumnInfo(name = ConfigsEntity.PAYMENT_COMMISSION) val paymentCommission: Double,
+    @ColumnInfo(name = ConfigsEntity.PAYMENT_COMMISSION) val paymentCommission: Float,
     @ColumnInfo(name = ConfigsEntity.SUPPORTED_CURRENCIES) val supportedCurrencies: CurrencyCachedList,
     @ColumnInfo(name = ConfigsEntity.SUPPORTED_DISTANCE_UNITS) val supportedDistanceUnits: StringList,
     @ColumnInfo(name = ConfigsEntity.OFFICE_PHONE) val officePhone: String,
